Extract cache_headers module to its own file
